Reorganizacion
Se movieron las rutas de config al middleware para un mayor orden, se conserva aun el config por si ocurre algun error
