+++ src/config/fileUploadConfig.js
@@ -1,81 +1,70 @@
-import multer from "multer";
-import path from "path";
-import { fileURLToPath } from "url";
-import fs from "fs";
+// import multer from "multer";
+// import path from "path";
+// import { fileURLToPath } from "url";
 
-// Para obtener la ruta del directorio actual
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
+// const __filename = fileURLToPath(import.meta.url);
+// const __dirname = path.dirname(__filename);
 
-// Configuración de almacenamiento de archivos generales
-const storage = multer.diskStorage({
-  destination: function (req, _file, cb) {
-    cb(null, path.join(__dirname, "../public/uploads"));
-  },
-  filename: function (req, file, cb) {
-    cb(null, `${Date.now()}-${file.originalname}`);
-  },
-});
+// const storage = multer.diskStorage({
+//   destination: function (req, _file, cb) {
+//     cb(null, path.join(__dirname, "../public/uploads"));
+//   },
+//   filename: function (req, file, cb) {
+//     cb(null, `${Date.now()}-${file.originalname}`);
+//   },
+// });
 
-const upload = multer({
-  storage,
-  limits: { fileSize: 20 * 1024 * 1024 },
-  fileFilter: (req, file, cb) => {
-    const allowedTypes = ['application/pdf', 'image/jpeg', 'image/png'];
-    if (!allowedTypes.includes(file.mimetype)) {
-      return cb(new Error('Tipo de archivo no permitido'), false);
-    }
-    cb(null, true);
-  },
-});
+// const upload = multer({
+//   storage,
+//   limits: { fileSize: 20 * 1024 * 1024 },
+//   fileFilter: (req, file, cb) => {
+//     const allowedTypes = ['application/pdf', 'image/jpeg', 'image/png'];
+//     if (!allowedTypes.includes(file.mimetype)) {
+//       return cb(new Error('Tipo de archivo no permitido'), false);
+//     }
+//     cb(null, true);
+//   },
+// });
 
-// Configuración de almacenamiento de archivos del usuario
-const storageUser = multer.diskStorage({
-  destination: function (req, _file, cb) {
-    cb(null, path.join(__dirname, "../public/users"));
-  },
-  filename: function (req, file, cb) {
-    cb(null, `${Date.now()}-${file.originalname}`);
-  },
-});
+// const storageUser = multer.diskStorage({
+//   destination: function (req, _file, cb) {
+//     cb(null, path.join(__dirname, "../public/users"));
+//   },
+//   filename: function (req, file, cb) {
+//     cb(null, `${Date.now()}-${file.originalname}`);
+//   },
+// });
 
-const uploadUser = multer({
-  storage: storageUser, limits: { fileSize: 20 * 1024 * 1024 },
-  fileFilter: (req, file, cb) => {
-    const allowedTypes = ['application/pdf', 'image/jpeg', 'image/png'];
-    if (!allowedTypes.includes(file.mimetype)) {
-      return cb(new Error('Tipo de archivo no permitido'), false);
-    }
-    cb(null, true);
-  },
-});
+// const uploadUser = multer({
+//   storage: storageUser, limits: { fileSize: 20 * 1024 * 1024 },
+//   fileFilter: (req, file, cb) => {
+//     const allowedTypes = ['application/pdf', 'image/jpeg', 'image/png'];
+//     if (!allowedTypes.includes(file.mimetype)) {
+//       return cb(new Error('Tipo de archivo no permitido'), false);
+//     }
+//     cb(null, true);
+//   },
+// });
 
-// Configuración de multer para los archivos de entregas
-const storageDeliveries = multer.diskStorage({
-  destination: function (req, _file, cb) {
-    const dir = path.join(__dirname, "../public/deliveries");
-    if (!fs.existsSync(dir)) {
-      fs.mkdirSync(dir, { recursive: true });
-    }
-    cb(null, dir);
-  },
-  filename: function (req, file, cb) {
-    cb(null, `${Date.now()}-${file.originalname}`);
-  },
-});
+// const storageDeliveries = multer.diskStorage({
+//   destination: function (req, _file, cb) {
+//     cb(null, path.join(__dirname, "../public/deliveries"));
+//   },
+//   filename: function (req, file, cb) {
+//     cb(null, `${Date.now()}-${file.originalname}`);
+//   },
+// });
 
-const uploadDelivery = multer({
-  storage: storageDeliveries,
-  limits: { fileSize: 20 * 1024 * 1024 },
-  fileFilter: (req, file, cb) => {
-    const allowedTypes = ['application/pdf', 'image/jpeg', 'image/png'];
-    if (!allowedTypes.includes(file.mimetype)) {
-      return cb(new Error('Tipo de archivo no permitido'), false);
-    }
-    cb(null, true);
-  },
-});
+// const uploadDelivery = multer({
+//   storage: storageDeliveries,
+//   limits: { fileSize: 20 * 1024 * 1024 },
+//   fileFilter: (req, file, cb) => {
+//     const allowedTypes = ['application/pdf', 'image/jpeg', 'image/png'];
+//     if (!allowedTypes.includes(file.mimetype)) {
+//       return cb(new Error('Tipo de archivo no permitido'), false);
+//     }
+//     cb(null, true);
+//   },
+// });
 
-
-
-export { upload, uploadUser, uploadDelivery };
+// export { upload, uploadUser, uploadDelivery };
